chore: pass renderProps for create alert button
